Blood Stain Prototype
Drawing on the wall freaks me out. I manage to make a simpler and
bloody effect that the player leaves blood stains on the ground. Which
can be as useful as a mark in the maze. And we can add some time
pressure since the role is bleeding.

Check out the scene BloodPrint.